requirements.txt
# Core Flask - Use older version compatible with Dash
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.4
Flask-CORS==4.0.0
Werkzeug==2.2.3

# Database
psycopg2-binary==2.9.7
SQLAlchemy-Utils==0.41.1

# Face Recognition & Image Processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
# dlib and face-recognition can be problematic on Windows - optional
# dlib==19.24.2
# face-recognition==1.3.0
# face-recognition-models==0.3.0

# Data Processing
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1

# HTTP & API
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.0.0

# Async & Background Tasks
celery==5.3.1
redis==4.6.0
hiredis==2.2.3
eventlet==0.33.3
python-socketio==5.9.0
APScheduler==3.10.4

# Web Server
gunicorn==21.2.0
gevent==24.2.1

# Notifications
twilio==8.5.0

# AWS & Cloud (optional)
boto3==1.28.17

# Monitoring & Debug
flower==2.0.0

# Security
pyjwt==2.8.0
bcrypt==4.0.1
cryptography==41.0.3

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Excel/CSV Export
xlwt==1.3.0
openpyxl==3.1.2
XlsxWriter==3.1.2
reportlab==4.0.4
WeasyPrint==60.1

# QR Code (optional)
qrcode==7.4.2
pyzbar==0.1.9